+++ build.sh
@@ -35,7 +35,7 @@ cat > "$STAGE/Contents/Info.plist" <<'PLIST'
   <key>CFBundleExecutable</key><string>Ambit</string>
   <key>CFBundleIdentifier</key><string>dev.ambit.Ambit</string>
   <key>CFBundlePackageType</key><string>APPL</string>
-  <key>CFBundleShortVersionString</key><string>1.0.1</string>
+  <key>CFBundleShortVersionString</key><string>1.1.0</string>
   <key>CFBundleVersion</key><string>1</string>
   <key>CFBundleIconFile</key><string>AppIcon</string>
   <key>CFBundleIconName</key><string>AppIcon</string>